+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/buttons/InlineKeyboardButtons/InlineKeyboardButton.kt
@@ -3,7 +3,7 @@ package dev.inmo.tgbotapi.types.buttons.InlineKeyboardButtons
 import dev.inmo.tgbotapi.types.*
 import dev.inmo.tgbotapi.types.games.CallbackGame
 import kotlinx.serialization.*
-import kotlinx.serialization.json.JsonElement
+import kotlinx.serialization.json.*
 
 /**
  * Some button of [dev.inmo.tgbotapi.types.buttons.InlineKeyboardMarkup]. See inheritors and visit
@@ -15,10 +15,14 @@ sealed interface InlineKeyboardButton {
 }
 
 @Serializable
-data class UnknownInlineKeyboardButton internal constructor(
-    override val text: String,
+data class UnknownInlineKeyboardButton (
     val rawData: JsonElement
-) : InlineKeyboardButton
+) : InlineKeyboardButton {
+    override val text: String
+        get() = runCatching {
+            rawData.jsonObject[textField] ?.jsonPrimitive ?.content
+        }.getOrNull() ?: ""
+}
 
 /**
  * This type of button must always be the first button in the first row. Visit